Move materials to ResourceManager from Scene
